Support mongoDb replica sets and local mongo _id
Use URIParser to create connections so that connections to either a
single mongo db or a replica set are supported. A replica set connection
is required in order to be able to keep inserting documents in the
replica set when the master fails over.

Add an option to automatically generate a mongo _id field using the
event timestamp when inserting a document.


Former-commit-id: e6c703839904730db304655ddac79b0497b99d18