[rabbitmq] Rearrange RabbitMQ input configuration attributes
Exchange and routing key make little sense in an input (which
is concerned with how to consume messages). As a result, they
will likely only cause confusion.

For backwards compatibility, they're made optional now.

Attributes were regrouped to reflect which of them are
related to the queue and which are connection attributes.

Finally, extra queue arguments (such as x-ha-policy) are
now used again.


Former-commit-id: 17a498e749da578aedd8f6c2dc20c88cde262073