inputs: rabbitmq: fix optional exchange binding
Make hot_bunnies implementation behave the same as the bunny
implementation: optionally bind the input queue to the output exchange
if the "exchange" option is defined.  If it is not defined, some other
out-of-band process will need to setup the bindings.  Remove the
backward-compatibility reference, because this is actually a useful
feature to continue supporting.

LOGSTASH-1300


Former-commit-id: 0c19e75d499cc2a1efacc126f200fa6e1c194123