Additional csv option for quote_char
This will allow users to configure a different quoting char in cases where the data may not exactly fit the csv standard.

A common example may be to completely ignore the quote char, and to do this a null \x00 char can be used instead.
Another example may be csv with ' as the quote char instead of "


Former-commit-id: 44a9bca63ee1ed0cc9a482f6dc57772de6f74aa1