Set the default codec to JSON
Sending structured messages is oh so nice! The JSON codec
sends a JSON representation of the Logstash event on output
and parses it on input. This allows you to more easily
send data between Logstash agents using Kafka as well as
easier integration with tools sending and receiving events
from Logstash over Kafka!


Former-commit-id: 7586a0b5db5fe6757440bb71106715e65113cff0